Allow anyone to batch_mint Only allow minting for yourself Organize tests
