+++ src/lib.rs
@@ -6,7 +6,10 @@ mod rand;
 pub mod receiver;
 pub mod state;
 pub mod token;
+mod unittest_batch_mint;
 mod unittest_handles;
+mod unittest_helpers;
+mod unittest_mint;
 mod unittest_queries;
 mod utils;
 mod viewing_key;
